pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conformer-tts"
version = "0.1.0"
description = "Conformer-based TTS for LegacyAI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "LegacyAI" }]
dependencies = [
    "torch>=2.1",
    "torchaudio>=2.1",
    "pytorch-lightning>=2.2",
    "numpy>=1.24",
    "scipy>=1.11",
    "librosa>=0.10",
    "soundfile>=0.12",
    "phonemizer>=3.2",
    "unidecode>=1.3",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "tensorboard>=2.15",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.3",
    "black>=24.0",
    "mypy>=1.8",
]

[tool.setuptools.packages.find]
include = ["conformer_tts*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
